requirements.txt
requests
discord.py
python-dotenv
nhentai.py==1.0.5
bs4
lxml
asyncio
img2pdf
Pillow
pytz
aiofiles
aiohttp
aiozipstream